+++ app/renderer/pages/wallet/account/AccountDetails.js
@@ -6,6 +6,7 @@ import { MainContent, MainLayout } from 'components/layout';
 import { Clipboard, PageHeading, PageFooter, Tabs, TabPane } from 'components';
 import PortfolioSection from './PortfolioSection';
 import ReceiveSection from './ReceiveSection';
+import TransferSection from './TransferSection';
 
 const Subheading = ({ account }) => {
   return (
@@ -16,21 +17,23 @@ const Subheading = ({ account }) => {
   );
 };
 
-const AccountDetails = ({ account }) => {
+const AccountDetails = ({ account, onTransfer, currentWallet }) => {
   return account ? (
     <React.Fragment>
       <PageHeading subHeading={<Subheading {...{ account }} />}>Account name here...</PageHeading>
       <div className="content">
-        <Tabs defaultActiveKey="2" styles={{ maxHeight: '730px' }}>
-          <TabPane tab="Portfolio" key="1">
+        <Tabs defaultActiveKey="1" >
+          <TabPane tab="Portfolio" key="1" withScrollable>
             <PortfolioSection {...{ account }} />
           </TabPane>
-          <TabPane tab="Receive" key="2">
+          <TabPane tab="Deposit" key="2" withScrollable>
             <ReceiveSection {...{ account }} />
+          </TabPane>
+          <TabPane tab="Send" key="3" >
+            <TransferSection {...{ account, onTransfer, currentWallet }} />
           </TabPane>
         </Tabs>
       </div>
-      <PageFooter blank />
     </React.Fragment>
   ) : (
     <PageHeading subHeading="There is no accessible account"> Wallet account </PageHeading>