+++ app/renderer/pages/wallet/account/index.js
@@ -10,13 +10,13 @@ import withContainer from './container';
 import WalletDetailsSubNav from './WalletDetailsSubNav';
 import AccountDetails from './AccountDetails';
 
-const WalletDetailsPage = ({ wallets, match }) => {
+const WalletDetailsPage = ({ wallets, match, onTransfer }) => {
   const { walletId, accountPublicAddress } = match.params;
   const wallet = wallets && R.find(R.propEq('id', walletId))(wallets);
   return wallet ? (
     <MainLayout subNav={<WalletDetailsSubNav currentWallet={wallet} {...{ wallets }} />}>
       <MainContent display="flex">
-        <AccountDetails account={wallet.accounts[accountPublicAddress]} />
+        <AccountDetails account={wallet.accounts[accountPublicAddress]} onTransfer={onTransfer} currentWallet={wallet} />
       </MainContent>
     </MainLayout>
   ) : (