+++ app/common/types/cennznet-node.types.js
@@ -21,6 +21,11 @@ export const NetworkNameMapping = {
   'Kauri CENNZnet': 'kauri',
 };
 
+export const CENNZScanTxUrl = {
+  'rimu': 'https://service.centrality.me/cennzscan/tx',
+  'kauri': 'https://service.centrality.me/cennzscan/tx', // TODO CENNZScan not support DEV yet
+};
+
 export const PreDefinedAssetIdObj = {
   STAKING_TOKEN: { ID: PreDefinedAssetId.stakingToken, NAME: PreDefinedAssetIdName[PreDefinedAssetId.stakingToken], BN: new BN(PreDefinedAssetId.stakingToken, 10)},
   SPENDING_TOKEN: { ID: PreDefinedAssetId.spendingToken, NAME: PreDefinedAssetIdName[PreDefinedAssetId.spendingToken], BN: new BN(PreDefinedAssetId.spendingToken, 10)},